requirements2.txt
alabaster==0.7.8
alembic==0.8.4
appnope==0.1.0
attrs==16.3.0
avro==1.8.1
Babel==2.3.4
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.4.0.2
backports.weakref==1.0rc1
BeautifulSoup==3.2.1
beautifulsoup4==4.4.1
bleach==1.5.0
botocore==1.4.81
bs4==0.0.1
cassandra-driver==3.6.0
catkin-pkg==0.2.10
Cerberus==0.9.2
certifi==2015.11.20.1
cffi==1.9.1
chainer==2.1.0
Chameleon==2.24
characteristic==14.3.0
chardet==2.3.0
click==6.6
configparser==3.5.0
constantly==15.1.0
contextlib2==0.5.3
cryptography==1.6
cssselect==1.0.0
cycler==0.10.0
D3epPyramid==0.0
daft==0.0.4
datalab==1.1.0
dataset==0.6.1
decorator==4.0.4
Django==1.9.7
django-bootstrap3==7.0.1
django-registration-redux==1.1
docutils==0.12
dominate==2.2.0
empy==3.3.2
entrypoints==0.2.2
enum34==1.1.6
Eve==0.6.1
Events==0.2.1
filelock==2.0.12
Flask==0.10.1
Flask-Bootstrap==3.1.1.2
Flask-Cors==2.1.2
Flask-Login==0.3.2
Flask-Moment==0.5.1
Flask-PyMongo==0.3.1
Flask-Restless==0.17.0
Flask-Script==0.6.7
Flask-SocketIO==2.7.1
Flask-SQLAlchemy==1.0
Flask-WTF==0.12
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==3.0.5
gapic-google-logging-v2==0.9.3
gapic-google-pubsub-v1==0.9.3
GDAL==1.11.5
gnureadline==6.3.3
google-api-python-client==1.6.2
google-cloud==0.19.0
google-gax==0.13.0
Google-Search-API==1.1.12
googleapis-common-protos==1.5.3
graphviz==0.4.10
grpc-google-logging-v2==0.9.3
grpc-google-pubsub-v1==0.9.3
grpcio==1.6.3
gspread==0.2.5
guppy==0.1.10
html5lib==0.9999999
httplib2==0.10.3
idna==2.1
imagesize==0.7.1
incremental==16.10.1
ipaddress==1.0.17
ipykernel==4.5.2
ipython==4.0.1
ipython-genutils==0.1.0
ipython-sql==0.3.8
ipywidgets==4.1.1
itsdangerous==0.24
jdcal==1.2
Jinja2==2.7.2
jmespath==0.9.0
jsonify==0.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==4.1.1
jupyter-console==4.0.3
jupyter-core==4.0.6
Keras==2.0.5
LEPL==5.1.3
line-profiler==1.0
livereload==2.5.1
lxml==3.6.4
Mako==1.0.3
Markdown==2.6.8
MarkupSafe==0.21
matplotlib==1.5.3
mecab-python==0.996
mimerender==0.5.5
mistune==0.7.1
mkdocs==0.16.3
mock==2.0.0
nbconvert==4.1.0
nbformat==4.0.1
ndg-httpsclient==0.4.0
nltk==3.2.5
normality==0.2.4
nose==1.3.7
notebook==4.0.6
numpy==1.11.2
oauth2client==2.2.0
oauthlib==1.0.3
pandas==0.17.1
pandas-profiling==1.4.0
parsel==1.1.0
PasteDeploy==1.5.2
path.py==8.1.2
pathlib2==2.1.0
pbr==1.9.1
pep8==1.7.0
pexpect==4.0.1
pickleshare==0.5
Pillow==3.4.2
plotly==1.12.5
ply==3.8
prettytable==0.7.2
protobuf==3.3.0
psutil==4.3.0
ptyprocess==0.5
py==1.4.31
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.17
PyDispatcher==2.0.5
pydot==1.1.0
pyflakes==1.3.0
Pygments==2.0.2
pymc==2.3.6
pymongo==2.9.1
PyMySQL==0.7.9
PyMySQL3==0.5
pyOpenSSL==16.2.0
pyowm==2.3.1
pyparsing==2.1.10
PyPrind==2.9.8
pyramid==1.7
pyramid-chameleon==0.3
pyramid-debugtoolbar==3.0.1
pyramid-jinja2==2.6.2
pyramid-mako==1.0.2
pyramid-tm==0.12.1
pyserial==3.2.1
PySide==1.2.4
pytest==2.8.7
python-dateutil==2.5.0
python-editor==0.5
python-engineio==1.0.3
python-mimeparse==0.1.4
python-socketio==1.5.1
pytz==2016.10
PyYAML==3.11
pyzmq==15.1.0
qtconsole==4.1.1
queuelib==1.4.2
redis==2.10.5
repoze.lru==0.6
requests==2.9.1
requests-cache==0.4.10
requests-oauthlib==0.5.0
rosdep==0.11.5
rosdistro==0.4.7
rosinstall==0.7.8
rosinstall-generator==0.1.11
rospkg==1.0.39
rsa==3.4.2
scikit-learn==0.18.2
scipy==0.17.0
Scrapy==1.2.1
seaborn==0.7.0
selenium==3.0.1
service-identity==16.0.0
simplegeneric==0.8.1
simplejson==3.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.4.5
SQLAlchemy==0.9.4
sqlparse==0.2.4
TBB==0.1
tensorflow==1.2.1
tensorflow-gpu==1.1.0
terminado==0.5
Theano==0.8.2
tornado==4.3
traitlets==4.3.2
transaction==1.6.1
translationstring==1.3
tweepy==3.5.0
Twisted==16.6.0
Unidecode==0.4.19
uritemplate==3.0.0
urllib3==1.19.1
vboxapi==1.0
vcrpy==1.7.4
vcstools==0.1.38
venusian==1.0
virtualenv==15.1.0
visitor==0.1.3
w3lib==1.16.0
waitress==0.9.0
WebOb==1.6.1
Werkzeug==0.9.4
widgetsnbextension==1.2.2
wrapt==1.10.8
wstool==0.1.13
WTForms==2.1
zope.deprecation==4.1.2
zope.interface==4.3.2
zope.sqlalchemy==0.7.6
